% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptar.R
\name{adaptar_travel_contact}
\alias{adaptar_travel_contact}
\title{Adaptar datos de Travel Contact a plantilla de alta de hospedaje}
\usage{
adaptar_travel_contact(
  travel_contact = "Travel Contact.xlsx",
  archivo = "tc_plantilla.xlsx"
)
}
\arguments{
\item{travel_contact}{Nombre del archivo Excel de entrada. Por defecto `"Travel Contact.xlsx"`.}

\item{archivo}{Nombre del archivo Excel de salida. Por defecto `"tc_plantilla.xlsx"`.}
}
\value{
Nombre del archivo generado.
}
\description{
Toma un archivo de Excel exportado desde Travel Contact, transforma los datos
al formato requerido por una plantilla de alta de reserva de hospedaje y los
inserta en el archivo de salida.
}
