% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fecha.R
\name{es_mayor_de_edad}
\alias{es_mayor_de_edad}
\title{Verifica si una persona es mayor de edad}
\usage{
es_mayor_de_edad(fecha_nacimiento)
}
\arguments{
\item{fecha_nacimiento}{Una cadena de texto con la fecha de nacimiento en formato "AAAA-MM-DD".}
}
\value{
`TRUE` si la persona tiene 18 a침os o m치s, `FALSE` en caso contrario.
}
\description{
Esta funci칩n determina si una persona es mayor de 18 a침os a partir de su fecha de nacimiento.
}
\keyword{internal}
