% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generar_xml.R
\name{generar_xml}
\alias{generar_xml}
\title{Generar partes de viajeros XML desde una hoja de cálculo}
\usage{
generar_xml(hoja_calculo, archivo_xml = NULL, pdf = TRUE, optimizar = FALSE)
}
\arguments{
\item{hoja_calculo}{Carácter. Ruta al archivo de hoja de cálculo que se va a procesar.}

\item{archivo_xml}{Carácter (opcional). Ruta del archivo XML de salida. Si es NULO,
el archivo XML tendrá el mismo nombre y ubicación que el archivo de hoja de cálculo.}

\item{pdf}{Booleano. Sigue el modelo del archivo pdf o del ejemplo xml.}

\item{optimizar}{Booleano. Elimina los nodos vacíos del archivo xml.}
}
\value{
Carácter. La ruta del archivo del documento XML generado.
}
\description{
Esta función lee un archivo de hoja de cálculo (Excel o ODS) y genera un archivo
de salida XML que contiene datos estructurados.
}
\examples{
hoja_calculo <- system.file("extdata", "pdf/partes_viajeros.xlsx", package = "comunicaXML")
archivo_xml <- tempfile(fileext = ".xml")

archivo <- generar_xml(hoja_calculo, archivo_xml)

}
