% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtener.R
\name{obtener_hoja_calculo}
\alias{obtener_hoja_calculo}
\title{Obtener una hoja de cálculo predefinida}
\usage{
obtener_hoja_calculo(archivo, pdf = TRUE)
}
\arguments{
\item{archivo}{Ruta del archivo de destino donde se copiará la hoja de cálculo.
Debe incluir la extensión `.xlsx` o `.ods`.}

\item{pdf}{Booleano. Sigue el modelo del archivo pdf o del ejemplo xml.}
}
\value{
Devuelve el nombre del archivo.
}
\description{
Esta función copia una hoja de cálculo predefinida a una ubicación especificada
por el usuario. Se admiten los formatos `.xlsx` y `.ods`.
}
\details{
Si el archivo de destino ya existe, la función detiene la ejecución con un error.
Si el formato del archivo no es compatible, también se genera un error.
}
\examples{
\dontrun{
obtener_hoja_calculo("mi_hoja.xlsx")
obtener_hoja_calculo("mi_hoja.ods")
}

}
