% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tipo_documento}
\alias{tipo_documento}
\title{Tipos de documentos}
\format{
A data frame of strings.
}
\usage{
tipo_documento
}
\description{
Códigos y descripciones de tipos de documentos.
}
\keyword{datasets}
