% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tipo_establecimiento}
\alias{tipo_establecimiento}
\title{Tipos de establecimientos}
\format{
A data frame of strings.
}
\usage{
tipo_establecimiento
}
\description{
Códigos y descripciones de tipos de establecimientos.
}
\keyword{datasets}
