% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dni.R
\name{validar_dni}
\alias{validar_dni}
\title{Validar y avisar de DNI español no válido}
\usage{
validar_dni(dni, ubicacion)
}
\arguments{
\item{dni}{Una cadena de caracteres que representa el DNI a validar.}

\item{ubicacion}{Una cadena de caracteres que indica la ubicación o contexto en el que se
se está realizando la validación. Esta información está incluida en el mensaje de advertencia.}
}
\value{
Siempre se devuelve `TRUE`, pero se emite un aviso si el DNI no es válido.
}
\description{
Comprueba si un determinado DNI español (Documento Nacional de Identidad)
es válido. Si el DNI no es válido se emite un aviso especificando la ubicación.
}
\keyword{internal}
