% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email.R
\name{validar_email}
\alias{validar_email}
\title{Validar una dirección de correo electrónico}
\usage{
validar_email(email, ubicacion)
}
\arguments{
\item{email}{Una cadena de caracteres que representa la dirección de correo electrónico a validar.}

\item{ubicacion}{Carácter. Una descripción de la ubicación donde se realiza la validación.}
}
\value{
Lógico. Siempre devuelve TRUE.
}
\description{
Comprueba si una dirección de correo electrónico determinada sigue un formato válido.
}
\keyword{internal}
