% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/establecimiento.R
\name{validar_establecimiento}
\alias{validar_establecimiento}
\title{Validar datos del establecimiento}
\usage{
validar_establecimiento(comunicacion, establecimiento)
}
\arguments{
\item{comunicacion}{Un data frame que representa la tabla de comunicación.
Debe contener una columna de clave principal.}

\item{establecimiento}{Un data frame que representa la tabla de establecimiento.
Debe contener los campos necesarios:
- `codigo`: Código del establecimiento.
- `pais`: Código de país (ISO 3).
- `codigoMunicipio`: Código de municipio (para España).
- `nombreMunicipio`: Nombre del municipio (para ubicaciones no españolas).}
}
\value{
`TRUE`. Genera advertencias si:
: los datos del establecimiento están incompletos o son inconsistentes.
: los detalles del municipio no coinciden con los requisitos del país.
: se proporcionan simultáneamente un código de establecimiento y detalles adicionales.
}
\description{
Comprueba si los datos del establecimiento proporcionados están estructurados correctamente y son coherentes.
Comprueba que:
: la clave externa de establecimiento existe en la tabla de comunicación.
: los campos obligatorios se completan según el país.
: la información del municipio está especificada correctamente según el país.
: se proporciona un código de establecimiento válido o detalles completos del establecimiento, pero no ambos.
}
\keyword{internal}
