% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fecha.R
\name{validar_fecha}
\alias{validar_fecha}
\title{Validar formato de fecha}
\usage{
validar_fecha(fecha, ubicacion)
}
\arguments{
\item{fecha}{Carácter. La cadena de fecha para validar.}

\item{ubicacion}{Carácter. Una descripción de la ubicación donde se realiza la validación.}
}
\value{
Lógico. Siempre devuelve TRUE.
}
\description{
Comprueba si una fecha determinada sigue el formato 'AAAA-MM-DD'.
Si el formato es incorrecto, se emite una advertencia con la ubicación proporcionada.
}
\keyword{internal}
