% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fecha.R
\name{validar_fecha_hora}
\alias{validar_fecha_hora}
\title{Validar formato de fecha y hora}
\usage{
validar_fecha_hora(fecha_hora, ubicacion)
}
\arguments{
\item{fecha_hora}{Carácter. La cadena de fecha para validar.}

\item{ubicacion}{Carácter. Una descripción de la ubicación donde se realiza la validación.}
}
\value{
Lógico. Siempre devuelve TRUE.
}
\description{
Comprueba si una fecha y hora determinada sigue el formato 'AAAA-MM-DDThh:mm:ss'.
Si el formato es incorrecto, se emite una advertencia con la ubicación proporcionada.
}
\keyword{internal}
