% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fecha.R
\name{validar_fecha_tarjeta}
\alias{validar_fecha_tarjeta}
\title{Validar fecha de vencimiento de la tarjeta de crédito}
\usage{
validar_fecha_tarjeta(fecha, ubicacion)
}
\arguments{
\item{fecha}{Carácter. La cadena de fecha de vencimiento para validar.}

\item{ubicacion}{Carácter. Una descripción de la ubicación donde se realiza la validación.}
}
\value{
Lógico. Siempre devuelve TRUE.
}
\description{
Comprueba si la fecha de vencimiento de una tarjeta de crédito determinada sigue
el formato 'mm/aaaa'. También verifica si la fecha es válida y no ha caducado.
}
\details{
Si se encuentra algún problema, se emite una advertencia con la ubicación proporcionada.
}
\keyword{internal}
