% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/municipio.R
\name{validar_municipio}
\alias{validar_municipio}
\title{Validar y advertir sobre un código de municipio no válido}
\usage{
validar_municipio(codigo_municipio, ubicacion)
}
\arguments{
\item{codigo_municipio}{Una cadena de caracteres que representa el código del
municipio a validar.}

\item{ubicacion}{Una cadena de caracteres que indica la ubicación o contexto
en el que se está realizando la validación. Esta información está incluida en
el mensaje de advertencia.}
}
\value{
`TRUE` siempre se devuelve, pero se emite una advertencia si el código d
e municipio no es válido.
}
\description{
Comprueba si un código de municipio determinado es válido.
Si el código de municipio no es válido, se emite una advertencia especificando
la ubicación.
}
\keyword{internal}
