% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dni.R
\name{validar_nif_nie}
\alias{validar_nif_nie}
\title{Validar y avisar de NIF o NIE no válido}
\usage{
validar_nif_nie(dni, ubicacion)
}
\arguments{
\item{dni}{Una cadena de caracteres que representa el NIF o NIE a validar.}

\item{ubicacion}{Una cadena de caracteres que indica la ubicación o contexto en el que se
se está realizando la validación. Esta información está incluida en el mensaje de advertencia.}
}
\value{
Siempre se devuelve `TRUE`, pero se emite un aviso si el NIF o NIE no es válido.
}
\description{
Comprueba si un determinado NIF o NIE es válido. Si no es válido se emite un
aviso especificando la ubicación.
}
\keyword{internal}
