% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/municipio.R
\name{validar_pais}
\alias{validar_pais}
\title{Validar y advertir sobre un código de país no válido}
\usage{
validar_pais(pais, ubicacion)
}
\arguments{
\item{pais}{Una cadena de caracteres que representa el código de país ISO de tres
letras a validar.}

\item{ubicacion}{Una cadena de caracteres que indica la ubicación o contexto en el
que se está realizando la validación. Esta información está incluida en el mensaje de
advertencia.}
}
\value{
`TRUE` siempre se devuelve, pero se emite una advertencia si el código de
país no es válido.
}
\description{
Comprueba si un código de país determinado es válido según la norma ISO 3166-1 alfa-3.
Si el código de país no es válido, se emite una advertencia especificando la ubicación.
}
\keyword{internal}
