% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persona.R
\name{validar_persona}
\alias{validar_persona}
\title{Validar datos de personas en la comunicación}
\usage{
validar_persona(comunicacion, persona)
}
\arguments{
\item{comunicacion}{Data frame con los datos de la comunicación.}

\item{persona}{Data frame con los datos de las personas asociadas a la comunicación.}
}
\value{
`TRUE` si todas las validaciones se completan sin problemas, de lo contrario, emite advertencias (`warning`).
}
\description{
Valida los datos de las personas asociadas a una comunicación, asegurando que
los campos obligatorios estén presentes y que la información proporcionada sea consistente.
Genera advertencias en caso de datos faltantes o incorrectos.
}
\details{
- Verifica que el campo `rol` solo tome los valores "TI" (Titular) o "VI" (Viajero).
- Asegura que al menos una persona tenga el rol de Titular (TI).
- Si hay menores de edad, al menos un adulto debe tener informado su parentesco con ellos.
- Valida la presencia de campos obligatorios como `nombre`, `apellido1` y `fechaNacimiento`.
- Verifica la existencia de `tipoDocumento` y `numeroDocumento`.
- Comprueba que `tipoDocumento` tenga un valor válido de `tipo_documento$codigo`.
- Valida que `sexo`, `nacionalidad` contengan valores válidos según las tablas de códigos.
- Se debe proporcionar al menos un dato de contacto: `telefono`, `telefono2` o `correo`.
- Se realizan validaciones sobre dirección, teléfono, email y documento de identidad.
}
\keyword{internal}
