% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solicitud.R
\name{validar_solicitud}
\alias{validar_solicitud}
\title{Validar una solicitud}
\usage{
validar_solicitud(solicitud)
}
\arguments{
\item{solicitud}{Un `data.frame` que contiene los datos de la solicitud.
Debe incluir al menos la columna `codigoEstablecimiento`.}
}
\value{
Devuelve `TRUE` si la validación se completa sin errores. Si falta
        algún campo requerido, genera advertencias (`warning`).
}
\description{
Esta función verifica que las solicitudes contengan todos los campos requeridos,
asegurando que los datos sean completos antes de su procesamiento.
}
\keyword{internal}
