% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telefono.R
\name{validar_telefono}
\alias{validar_telefono}
\title{Validar un teléfono}
\usage{
validar_telefono(telefono, ubicacion)
}
\arguments{
\item{telefono}{Una cadena de caracteres que representa el teléfono a validar.}

\item{ubicacion}{Carácter. Una descripción de la ubicación donde se realiza la validación.}
}
\value{
`TRUE` si el correo electrónico es válido, `FALSE` en caso contrario.
}
\description{
Comprueba si un teléfono determinado sigue un formato válido.
}
\keyword{internal}
