% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys.R
\name{validate_fk}
\alias{validate_fk}
\title{Validar columna de clave externa}
\usage{
validate_fk(table_pk, name_pk, table_fk, name_fk)
}
\arguments{
\item{table_pk}{Un data frame que representa la tabla que contiene la clave principal.}

\item{name_pk}{Una cadena de caracteres que indica el nombre de la tabla a la que se hace
referencia. La columna de clave principal se espera que tenga el nombre `{name_pk}_pk`.}

\item{table_fk}{Un data frame que representa la tabla que contiene la clave externa.}

\item{name_fk}{Una cadena de caracteres que indica el nombre de la tabla de referencia.
La columna de clave externa se espera que tenga el nombre `{name_pk}_fk`.}
}
\value{
`TRUE`. Genera advertencias si falta la columna de clave externa en la
tabla de referencia o si contiene valores que no están presentes en la columna de
clave principal de la tabla a la que se hace referencia.
}
\description{
Comprueba si existe una columna de clave externa en la tabla de referencia y garantiza
que todos los valores de la columna de clave externa coincidan con los valores existentes
en la columna de clave principal de la tabla a la que se hace referencia.
}
\keyword{internal}
