% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys.R
\name{validate_pk}
\alias{validate_pk}
\title{Validar columna de clave principal}
\usage{
validate_pk(table, name)
}
\arguments{
\item{table}{Un data frame que representa la tabla a validar.}

\item{name}{Una cadena de caracteres que indica el nombre de la tabla. La columna
de clave principal Se espera que tenga el nombre `{name}_pk`.}
}
\value{
`TRUE`. Genera advertencias si falta la columna de clave principal, contiene valores "NA",
o tiene duplicados.
}
\description{
Comprueba si una tabla contiene una columna de clave primaria con el formato correcto.
La columna de clave principal debe existir, no debe faltar ningún valor ("NA")
y contener valores únicos.
}
\keyword{internal}
